pubilc classes and restructuring
